back/requirements.txt
fastapi
uvicorn
pyrebase4
python-dotenv
pydantic[email]